requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
distlib==0.3.8
Django==5.0.3
django-bootstrap-datepicker-plus==5.0.5
django-countries==7.5.1
django-location-field==2.7.3
filelock==3.13.1
pillow==10.2.0
platformdirs==4.2.0
psycopg==3.1.18
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
